refactor(osmPlanet): analytics service
Split up service into multiple ones to increase readablity
