+++ osmPlanet/src/main/java/com/simra/konsumgandalf/osmPlanet/controller/AnalyticsController.java
@@ -1,10 +1,12 @@
 package com.simra.konsumgandalf.osmPlanet.controller;
 
+import com.simra.konsumgandalf.osmPlanet.services.AnalyticsServiceRegionMetrics;
+import com.simra.konsumgandalf.osmPlanet.services.AnalyticsServiceSimraRegionMetrics;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import com.simra.konsumgandalf.osmPlanet.services.AnalyticsService;
+import com.simra.konsumgandalf.osmPlanet.services.AnalyticsServiceHighwayMetrics;
 
 /**
  * Temporary controller to update the highway information
@@ -14,21 +16,27 @@ import com.simra.konsumgandalf.osmPlanet.services.AnalyticsService;
 public class AnalyticsController {
 
 	@Autowired
-	private AnalyticsService analyticsService;
+	private AnalyticsServiceHighwayMetrics analyticsServiceHighwayMetrics;
+
+	@Autowired
+	private AnalyticsServiceRegionMetrics analyticsServiceRegionMetrics;
+
+	@Autowired
+	private AnalyticsServiceSimraRegionMetrics analyticsServiceSimraRegionMetrics;
 
 	@PostMapping("/update/highway")
 	public void updateHighwayInformation() {
-		analyticsService.updateSafetyMetricsHighway();
+		analyticsServiceHighwayMetrics.updateSafetyMetricsHighway();
 	}
 
 	@PostMapping("/update/region")
 	public void updateRegionSafetyMetrics() {
-		analyticsService.calculateSafetyMetricsRegion();
+		analyticsServiceRegionMetrics.calculateSafetyMetricsRegion();
 	}
 
 	@PostMapping("/update/simra-region")
 	public void updateSimraRegionSafetyMetrics() {
-		analyticsService.calculateSafetyMetricsSimraRegion();
+		analyticsServiceSimraRegionMetrics.calculateSafetyMetricsSimraRegion();
 	}
 
 }
